+++ src/old/creep.ts
@@ -1,5 +1,5 @@
 import { StructureFilter, room_link } from "../linted/utility"
-import { Squad } from "classes/squad/squad"
+import { Squad } from "old/squad/squad"
 import { ChargeTarget } from "./room";
 
 export enum CreepStatus {  // @todo: add "meta" info to status and keep it on memory, to not change objectives between ticks
