+++ src/old/region.ts
@@ -1,9 +1,9 @@
 import { ErrorMapper } from "utils/ErrorMapper"
-import { Squad, SquadType, SquadMemory, SpawnPriority } from "classes/squad/squad"
-import { WorkerSquad } from "classes/squad/worker"
-import { ManualSquad } from "classes/squad/manual"
+import { Squad, SquadType, SquadMemory, SpawnPriority } from "old/squad/squad"
+import { WorkerSquad } from "old/squad/worker"
+import { ManualSquad } from "old/squad/manual"
 import { HarvesterSquad, HarvesterSquadMemory } from "./squad/harvester"
-import { ScoutSquad } from "classes/squad/scout"
+import { ScoutSquad } from "old/squad/scout"
 import { ActionResult } from "./creep"
 import { AttackerSquad } from "./squad/attacker"
 import { UpgraderSquad } from "./squad/upgrader";
