+++ src/old/room_layout.ts
