+++ src/old/squad/farmer.ts
@@ -1,6 +1,6 @@
 import { UID, room_link } from "../../linted/utility"
 import { Squad, SquadType, SquadMemory, SpawnPriority, SpawnFunction } from "./squad"
-import { CreepStatus, ActionResult, CreepType } from "classes/creep"
+import { CreepStatus, ActionResult, CreepType } from "old/creep"
 import { runTowers } from "../tower";
 import { ErrorMapper } from "../../utils/ErrorMapper";
