+++ src/old/squad/nuker_charger.ts
@@ -1,6 +1,6 @@
 import { UID, room_link } from "../../linted/utility"
 import { Squad, SquadType, SquadMemory, SpawnPriority, SpawnFunction } from "./squad"
-import { CreepStatus, ActionResult, CreepType } from "classes/creep"
+import { CreepStatus, ActionResult, CreepType } from "old/creep"
 
 export interface NukerChargerSquadMemory extends SquadMemory {
   nuker_id: string | undefined
