+++ src/old/squad/remote_harvester.ts
@@ -1,6 +1,6 @@
 import { UID, room_link } from "../../linted/utility"
 import { Squad, SquadType, SquadMemory, SpawnPriority, SpawnFunction, SquadStatus } from "./squad"
-import { CreepStatus, ActionResult, CreepType, CreepSearchAndDestroyOption } from "classes/creep"
+import { CreepStatus, ActionResult, CreepType, CreepSearchAndDestroyOption } from "old/creep"
 import { runHarvester } from "./harvester"
 import { Region } from "../region";
