+++ src/old/squad/remote_m_harvester.ts
@@ -1,6 +1,6 @@
 import { UID } from "../../linted/utility"
 import { Squad, SquadType, SquadMemory, SpawnPriority, SpawnFunction } from "./squad"
-import { CreepStatus, ActionResult, CreepType } from "classes/creep"
+import { CreepStatus, ActionResult, CreepType } from "old/creep"
 
 export interface RemoteMineralHarvesterSquadMemory extends SquadMemory {
   room_name: string
