+++ src/old/squad/worker.ts
@@ -1,6 +1,6 @@
 import { UID } from "../../linted/utility"
 import { Squad, SquadType, SquadMemory, SpawnPriority, SpawnFunction } from "./squad"
-import { CreepStatus, ActionResult, CreepType, WorkerSource } from "classes/creep"
+import { CreepStatus, ActionResult, CreepType, WorkerSource } from "old/creep"
 import { isNewLine } from "../../../node_modules/@types/acorn/index";
 
 interface WorkerSquadMemory extends SquadMemory {
